requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-timezone-field==6.1.0
django-tinymce==3.6.1
idna==3.6
kombu==5.3.4
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tpqoa==0.0.56
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.1.0
v20==3.0.25.0
vine==5.1.0
wcwidth==0.2.13
